requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
distlib==0.4.0
Django==5.2.3
djangorestframework==3.16.0
fastapi==0.116.1
filelock==3.18.0
h11==0.16.0
idna==3.10
joblib==1.4.2
mysql-connector-python==9.3.0
mysqlclient==2.2.7
narwhals==2.1.2
nltk==3.9.1
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0
virtualenv==20.33.0
